requirements.txt
aiohttp==3.7.2
asgiref==3.3.0
async-timeout==3.0.1
attrs==20.3.0
certifi==2020.6.20
chardet==3.0.4
Django==3.1.3
djangorestframework==3.12.2
flake8==3.8.4
idna==2.10
mccabe==0.6.1
multidict==5.0.0
pycodestyle==2.6.0
pyflakes==2.2.0
pytz==2020.4
requests==2.24.0
sqlparse==0.4.1
typing-extensions==3.7.4.3
urllib3==1.25.11
yarl==1.6.2
